SocialBot Class Refactor and frontend skeleton (#2)
SocialBot Class Refactor and frontend skeleton

* Set up React + Flask

Here, I was able to set up a rudimentary connection between the React frontend and the Flask backend. I made a temporary file called tester.py as a proof of concept, and will translate this into socialBot.py at the next commit.

* Rudimentary Frontend completed

The frontend is able to convey information to the backend, and the backend is able to convey information to the frontend. The social situation bot is able to display it's new responses to the website based on user input on the website

* Add Poetry, Update API key

* Create pyproject.toml

* Class Structure

Changed the structure of the code from random functions to a consolidated class structure so that we can easily create user accounts and user data. This prevents the muddling of overall state and prevents problems in the future with data.

* Fixed Errors

Changed the Flask portion to work with the new class structure that was implemented in the previous commit.

---------

Co-authored-by: Svar <spandya@imsa.edu>